test(db): extract cross-mode single-active tests to own file + review fixups (#1035, #1018)
Addresses post-commit review of 35fa1fd2:
- code-reviewer WARNING: rpc-single-active-session grew to 668L (>500 test
  ceiling). Revert it to master and move the cross-mode start-guard tests
  (Gap A internal_exam, Gap B vfr_rt) to a new self-contained file
  rpc-single-active-cross-mode.integration.test.ts (419L, own fixtures).
- red-team: add Gap B-3 (start_vfr_rt_exam_session auto-clears an abandoned
  discovery session), completing the discovery auto-clear coverage.
- code-reviewer WARNING: comment the two `!` non-null assertions in the
  emailed-resend test (safe via the preceding {error} null-check).
- semantic-reviewer SUGGESTION: expand the resend timestamp comment.
- doc-updater DRIFT: bump docs/plan.md integration count 305 -> 311 (+6).

25 integration tests pass locally across the 3 files (9 + 11 + 5).
