+++ packages/db/src/__integration__/rpc-single-active-session.integration.test.ts
@@ -10,13 +10,6 @@ import {
   getAnonClient,
   getAuthenticatedClient,
 } from './setup'
-import {
-  ensureBank,
-  getRtRefs,
-  insertDialogFillQuestion,
-  insertMcQuestion,
-  insertShortAnswerQuestion,
-} from './vfr-rt-helpers'
 
 // Single-active-session invariant (#1011, migs 136–142): a student may hold AT
 // MOST ONE active (ended_at IS NULL AND deleted_at IS NULL) quiz_sessions row
@@ -50,32 +43,6 @@ describe('RPC: start_discovery_session + single-active-session guard', () => {
   let refs: Awaited<ReturnType<typeof seedReferenceData>>
   const userIds: string[] = []
   const suffix = Date.now()
-  // internal_exam_codes rows created in Gap-A tests — hard-deleted in afterAll
-  // before cleanupTestData (they FK into users/orgs/quiz_sessions; no ON DELETE CASCADE).
-  const createdInternalExamCodeIds: string[] = []
-
-  // Insert a valid internal_exam_codes row for the test student via service-role
-  // (bypasses the no-INSERT RLS policy on that table, same as seedCode in
-  // rpc-record-internal-exam-code-emailed.integration.test.ts).
-  const seedCode = async (): Promise<{ id: string; code: string }> => {
-    const code = `SAS${suffix}${createdInternalExamCodeIds.length}`
-    const { data, error } = await admin
-      .from('internal_exam_codes')
-      .insert({
-        code,
-        subject_id: subjectId,
-        student_id: studentId,
-        issued_by: adminUserId,
-        organization_id: orgId,
-        expires_at: new Date(Date.now() + 24 * 60 * 60 * 1000).toISOString(),
-      })
-      .select('id')
-      .single()
-    if (error || !data) throw new Error(`seedCode: ${error?.message}`)
-    const id = data.id as string
-    createdInternalExamCodeIds.push(id)
-    return { id, code }
-  }
 
   // Insert an active session of an arbitrary mode for the test student directly
   // (service-role bypasses RLS, but NOT the partial unique index). Returns the id.
@@ -193,45 +160,8 @@ describe('RPC: start_discovery_session + single-active-session guard', () => {
   })
 
   afterAll(async () => {
-    const errors: string[] = []
-
-    // Step 1: hard-delete any internal_exam_codes rows created in Gap-A tests.
-    // These FK into users + organizations + quiz_sessions (no ON DELETE CASCADE)
-    // that cleanupTestData hard-deletes in step 2 — codes must be removed first
-    // or the FK constraint blocks user/org deletion (code-style.md §7 dependent steps).
-    if (createdInternalExamCodeIds.length > 0) {
-      try {
-        const { data: removed, error } = await admin
-          .from('internal_exam_codes')
-          .delete()
-          .in('id', createdInternalExamCodeIds)
-          .select('id')
-        if (error) throw new Error(error.message)
-        if ((removed?.length ?? 0) > 0) {
-          console.log(`[single-active-session] removed ${removed?.length} code(s)`)
-        }
-      } catch (e) {
-        errors.push(`code delete: ${e instanceof Error ? e.message : String(e)}`)
-      } finally {
-        createdInternalExamCodeIds.length = 0
-      }
-    }
-
-    // Steps 2-3 are FK-dependent on step 1 completing cleanly.
-    if (errors.length === 0) {
-      try {
-        await cleanupTestData({ admin, orgId, userIds })
-      } catch (e) {
-        errors.push(`cleanupTestData: ${e instanceof Error ? e.message : String(e)}`)
-      }
-      try {
-        await cleanupReferenceData({ admin, refs: [refs] })
-      } catch (e) {
-        errors.push(`cleanupReferenceData: ${e instanceof Error ? e.message : String(e)}`)
-      }
-    }
-
-    if (errors.length > 0) throw new Error(`afterAll: ${errors.join('; ')}`)
+    await cleanupTestData({ admin, orgId, userIds })
+    await cleanupReferenceData({ admin, refs: [refs] })
   })
 
   it('creates exactly one active discovery session and a discovery.started audit event', async () => {
@@ -484,185 +414,5 @@ describe('RPC: start_discovery_session + single-active-session guard', () => {
     expect(error).not.toBeNull()
     expect(error?.code).toBe('23505')
     expect(error?.message ?? '').toMatch(/uq_one_active_session_per_student/i)
-  })
-
-  // ── Gap A: start_internal_exam_session — single-active-session guard (mig 139) ──
-
-  it('blocks an internal exam start while another session is already active', async () => {
-    const { code } = await seedCode()
-    const quizId = await seedActiveSession('quick_quiz')
-    // Non-vacuity: the blocking session genuinely exists before the internal exam attempt.
-    expect((await readActiveSessions()).map((s) => s.id)).toContain(quizId)
-
-    const { data, error } = await studentClient.rpc('start_internal_exam_session', {
-      p_code: code,
-    })
-    expect(error).not.toBeNull()
-    expect(error?.message ?? '').toMatch(/another_session_active/i)
-    expect(data).toBeNull()
-
-    // No internal_exam session was created — the quick_quiz remains the only active session.
-    const after = await readActiveSessions()
-    expect(after.map((s) => s.id)).toEqual([quizId])
-    expect(after.map((s) => s.mode)).not.toContain('internal_exam')
-  })
-
-  it('auto-clears an abandoned discovery session when an internal exam starts', async () => {
-    const { code } = await seedCode()
-    const discoveryId = await seedActiveSession('discovery')
-    // Non-vacuity: the discovery row is genuinely active before the internal exam start.
-    expect((await readActiveSessions()).map((s) => s.id)).toEqual([discoveryId])
-
-    const { data, error } = await studentClient.rpc('start_internal_exam_session', {
-      p_code: code,
-    })
-    expect(error).toBeNull()
-    // start_internal_exam_session returns RETURNS TABLE — PostgREST serialises it as a
-    // one-row array. A non-empty array confirms the session was created.
-    const startedRows = requireRpcRows<{ session_id: string }>(
-      data,
-      'start_internal_exam_session (gap A-2)',
-    )
-    expect(startedRows.length).toBeGreaterThan(0)
-
-    // The discovery row is soft-deleted (not ended); the new internal_exam is the
-    // only active session.
-    const { data: discRow, error: discErr } = await admin
-      .from('quiz_sessions')
-      .select('ended_at, deleted_at')
-      .eq('id', discoveryId)
-      .single()
-    expect(discErr).toBeNull()
-    expect(discRow?.deleted_at).not.toBeNull()
-    expect(discRow?.ended_at).toBeNull()
-
-    const active = await readActiveSessions()
-    expect(active).toHaveLength(1)
-    expect(active[0]?.mode).toBe('internal_exam')
-  })
-
-  // ── Gap B: start_vfr_rt_exam_session — single-active-session guard (mig 140) ──
-
-  describe('start_vfr_rt_exam_session — single-active-session guard and idempotent resume', () => {
-    let rtSubjectId: string
-    let p1TopicId: string
-    let p2TopicId: string
-    let p3TopicId: string
-
-    beforeAll(async () => {
-      // Resolve the global RT subject/topics (seeded by mig 097; never deleted by tests).
-      const rtRefs = await getRtRefs()
-      rtSubjectId = rtRefs.rtSubjectId
-      p1TopicId = rtRefs.p1TopicId
-      p2TopicId = rtRefs.p2TopicId
-      p3TopicId = rtRefs.p3TopicId
-
-      // Seed the minimum question pool matching the per-part defaults used by
-      // start_vfr_rt_exam_session when parts_config is NULL (8 SA + 9 DF + 8 MC).
-      // ensureBank is idempotent — reuses an existing bank for the org if one
-      // was created by another describe in this file.
-      const bankId = await ensureBank(orgId, adminUserId)
-      for (let i = 0; i < 8; i++) {
-        await insertShortAnswerQuestion({
-          orgId,
-          bankId,
-          adminId: adminUserId,
-          rtSubjectId,
-          p1TopicId,
-          idx: 200 + i,
-        })
-      }
-      for (let i = 0; i < 9; i++) {
-        await insertDialogFillQuestion({
-          orgId,
-          bankId,
-          adminId: adminUserId,
-          rtSubjectId,
-          p2TopicId,
-          idx: 200 + i,
-        })
-      }
-      for (let i = 0; i < 8; i++) {
-        await insertMcQuestion({
-          orgId,
-          bankId,
-          adminId: adminUserId,
-          rtSubjectId,
-          p3TopicId,
-          idx: 200 + i,
-        })
-      }
-
-      // Seed an enabled exam_configs row for the RT subject in this test org.
-      // parts_config is omitted — the RPC falls back to the briefing-package
-      // defaults (P1=8 SA, P2=9 DF, P3=8 MC) via COALESCE. This is a different
-      // subject from the SAS exam_config seeded in the outer beforeAll, so the
-      // uq_exam_configs_org_subject_active partial unique index is not violated.
-      const { error: ecErr } = await admin.from('exam_configs').insert({
-        organization_id: orgId,
-        subject_id: rtSubjectId,
-        enabled: true,
-        total_questions: 25,
-        time_limit_seconds: 1800,
-        pass_mark: 75,
-      })
-      if (ecErr) throw new Error(`RT exam_configs insert: ${ecErr.message}`)
-    })
-
-    // No afterAll needed: the outer describe's afterAll calls cleanupTestData which
-    // removes all questions and exam_configs under orgId, including the RT rows
-    // seeded in this nested beforeAll.
-
-    it('blocks a new VFR RT session start while a different active session exists', async () => {
-      const quizId = await seedActiveSession('quick_quiz')
-      // Non-vacuity: the blocking session genuinely exists before the VFR RT start attempt.
-      expect((await readActiveSessions()).map((s) => s.id)).toContain(quizId)
-
-      const { data, error } = await studentClient.rpc('start_vfr_rt_exam_session', {
-        p_subject_id: rtSubjectId,
-      })
-      expect(error).not.toBeNull()
-      expect(error?.message ?? '').toMatch(/another_session_active/i)
-      expect(data).toBeNull()
-
-      // No vfr_rt_exam session was created — the quick_quiz remains the only active session.
-      const active = await readActiveSessions()
-      expect(active.map((s) => s.id)).toEqual([quizId])
-      expect(active.map((s) => s.mode)).not.toContain('vfr_rt_exam')
-    })
-
-    it('resumes the same VFR RT exam session on a second start call for the same subject', async () => {
-      // First call — no active session exists, so the RPC creates one.
-      const { data: firstData, error: err1 } = await studentClient.rpc(
-        'start_vfr_rt_exam_session',
-        { p_subject_id: rtSubjectId },
-      )
-      expect(err1).toBeNull()
-      const first = requireRpcResult<{ session_id: string; question_ids: string[] }>(
-        firstData,
-        'start_vfr_rt_exam_session (first call)',
-      )
-      expect(typeof first.session_id).toBe('string')
-      // Non-vacuity: the session genuinely exists as the only active one.
-      const activeAfterFirst = await readActiveSessions()
-      expect(activeAfterFirst.map((s) => s.id)).toContain(first.session_id)
-
-      // Second call with the same subject — the idempotent-resume path returns the
-      // existing session instead of raising another_session_active.
-      const { data: secondData, error: err2 } = await studentClient.rpc(
-        'start_vfr_rt_exam_session',
-        { p_subject_id: rtSubjectId },
-      )
-      expect(err2).toBeNull()
-      const second = requireRpcResult<{ session_id: string; question_ids: string[] }>(
-        secondData,
-        'start_vfr_rt_exam_session (second call)',
-      )
-
-      // Same session id and same frozen question set.
-      expect(second.session_id).toBe(first.session_id)
-      expect(Array.isArray(second.question_ids)).toBe(true)
-      expect(second.question_ids).toEqual(first.question_ids)
-    })
   })
 })
